+++ presets/react.js
@@ -1,7 +1,5 @@
 import js from '@eslint/js';
-import path from 'path';
 import tseslint from 'typescript-eslint';
-import { fileURLToPath } from 'url';
 
 import { browserGlobals } from '../rules/base/globals.js';
 import importConfig from '../rules/base/imports.js';
@@ -10,8 +8,6 @@ import tsConfig from '../rules/typescript/ts.js';
 import reactConfig from '../rules/react/react.js';
 import a11yConfig from '../rules/accessibility/a11y.js';
 import prettierConfig from '../rules/formatting/prettier.js';
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 export default tseslint.config([
 	{
@@ -48,9 +44,6 @@ export default tseslint.config([
 			},
 		},
 		settings: {
-			'better-tailwindcss': {
-				entryPoint: path.resolve(__dirname, '..', 'src/styles/tailwind.css'),
-			},
 			...reactConfig.settings,
 			...importConfig.settings,
 		},
@@ -80,9 +73,6 @@ export default tseslint.config([
 			},
 		},
 		settings: {
-			'better-tailwindcss': {
-				entryPoint: path.resolve(__dirname, '..', 'src/styles/tailwind.css'),
-			},
 			...reactConfig.settings,
 			...importConfig.settings,
 		},